restructure toplevel folders
