+++ libs/server/progression-engine/src/index.ts
@@ -1,18 +1,18 @@
 // Unified Services
-export * from './lib/core/events/event-processor.service';
-export * from './lib/core/events/event-publisher.service';
-export * from './lib/core/progression/user-progression.service';
+export * from './lib/events/event-processor.service';
+export * from './lib/events/event-publisher.service';
+export * from './lib/progression/services/user-progression.service';
 export * from './lib/core/activity/activity.service';
 
 // Supporting Services
-export * from './lib/core/services/badge.service';
-export * from './lib/core/services/level.service';
-export * from './lib/core/services/reward.service';
+export * from './lib/rewards/services/badge.service';
+export * from './lib/rewards/services/level.service';
+export * from './lib/rewards/services/reward.service';
 
 // Constants
-export * from './lib/constants/xp-values.config';
-export * from './lib/constants/level-thresholds';
-export * from './lib/constants/level-definitions.config';
+export * from './lib/config/xp-values.config';
+export * from './lib/config/level-thresholds';
+export * from './lib/config/level-definitions.config';
 
 // Utilities
 export * from './lib/utils/time-periods.utils';
