+++ libs/server/progression-engine/src/lib/actions/code-review/code-review-submit.handler.ts
@@ -1,6 +1,6 @@
 import { GameActionType } from '@codeheroes/types';
+import { XP_VALUES } from '../../config/xp-values.config';
 import { AbstractActionHandler } from '../base/abstract-action.handler';
-import { XP_VALUES } from '../../constants/xp-values.config';
 
 export class CodeReviewSubmitHandler extends AbstractActionHandler {
   protected actionType: GameActionType = 'code_review_submit';
