+++ libs/server/progression-engine/src/lib/config/level-definitions.config.ts
