+++ libs/server/progression-engine/src/lib/config/xp-values.config.ts
