--- libs/server/progression-engine/src/lib/core/events/index.ts
@@ -1,2 +1,0 @@
-export * from './event-processor.service';
-export * from './event-publisher.service';
