--- libs/server/progression-engine/src/lib/core/services/index.ts
@@ -1,3 +1,0 @@
-export * from './badge.service';
-export * from './level.service';
-export * from './reward.service';
