+++ libs/server/progression-engine/src/lib/events/event-processor.service.ts
@@ -2,7 +2,7 @@ import { DatabaseInstance, logger } from '@codeheroes/common';
 import { NotificationService } from '@codeheroes/notifications';
 import { Collections, ProgressionEvent, ProgressionEventType } from '@codeheroes/types';
 import { Firestore } from 'firebase-admin/firestore';
-import { BadgeService } from '../services/badge.service';
+import { BadgeService } from '../rewards/services/badge.service';
 
 export class EventProcessorService {
   private db: Firestore;
